fix : 카테고리 엔티티 필드 타입 변경
(order, depth) Long -> Integer

+++ src/main/java/site/caboomlog/backendservice/category/dto/CategoryResponse.java
@@ -16,7 +16,7 @@ public class CategoryResponse {
     private Long categoryId;
     private String categoryName;
     private Boolean categoryPublic;
-    private Long depth;
+    private Integer depth;
     private String topicName;
     private List<CategoryResponse> children = new ArrayList<>();
